Add README and LICENSE
PowerCrystals has given permission to license this project under the
MIT license and this commit adds documentation to reflect that.

Additionally, as this repository is no longer maintained, a readme has
been added to redirect visitors to skyboy's fork which is, as of
writing, the most active of the available forks.

See https://github.com/powercrystals/MineFactoryReloaded/pull/341